+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Stop injecting screens in `TheActivity` via the static component
 - Add support for sending teleconsult request via SMS
 - Show medicines required error in `TeleconsultPrescriptionScreen`
+- Remove `TeleconsultationApi` & load medical officers from `MedicalOfficer` table
 
 ### Fixes
 - Fix deeplink screens & warning dialogs displaying again after activity restart.